Note: Fun with branches on different machines.
I did LPtHW ex7.py on a Chromebook, in Vim in Termux, but could
not push bc OpenSSH was not set up correctly.  Working to resolve
that (need "external storage" set up).  Meanwhile, I want to do
more LPtHW on the laptop, but don't want to re-type ex7.py.  So,
skipping the Exercise 7 Study Drills for now, and working on
ex8.py.  This is good practice at managing multiple branches on
different machines.  This commit is just a note about this
situation, in the README.md and learning-git files.