PtHW ex13: argv (args to vars via commandline)
Intro to importing so-called features from a module.  I biffed
this at first, because I didn't realize I had to type additional
arguments on the command line after  python ex13.py
So, is argv a  function  or a  method  or what?  I think it is
a function, but I am not yet sure.